backend/pyproject.toml
[project]
name = "agent-factory-v4"
version = "0.1.0"
description = "AI-powered social media content creation platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.4.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "langsmith>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "google-genai>=1.0.0",
    "PyJWT>=2.9.0",
    "Pillow>=10.0.0",
    "colorthief>=0.2.1",
    "python-multipart>=0.0.12",
    "sse-starlette>=2.0.0",
    "aiofiles>=24.0.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "aiosqlite>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "agents", "brand"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
